==> scripts/10_run_in_stan.Rmd <==
 ---
title: "Stan model"
output: html_notebook
---
Stan version of model

# Read in data

```{r}
library(tidyverse)
library(forcats)

drugs_select <- "Airways"
source('02_arrange_data_arrays.R') 

## Select the first simulation
res <- res_list[[1]]
res <- res %>% 
  filter(drug_group_allocation == drugs_select)
rm(res_list, res_rag)

coef <- map(res$con, ~ .x$coef)
covar <- map(res$con, ~ .x$vcov)


n_studies <- nrow(res)

## Convert to matrix and array
covar_array <- array(NA, dim = c(n_studies, 10, 10))
coef_array <- array(NA, dim = c(n_studies,10))
for (j in 1:n_studies){
  covar_array[j, ,] <- covar[[j]]
  coef_array[j,]    <- coef[[j]] 
}
dim(coef_array)
dim(covar_array)

my_drug_n <- as.numeric(as_factor(res$my_drug))
study_id_n <- as.numeric(as_factor(as.character(res$study_id)))

## Add drug class data
drug_mtrx <- model.matrix(~ -1 + my_drug, res)

## Select depression:treatment interaction
coef_vect <- map_dbl(coef, ~ .x["dep:alloc"])
var_vect <- map_dbl(covar, ~ .x["dep:alloc", "dep:alloc"])

```

## Set-up stan
```{r}
library(rstan)
rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())
```

# Run example model, simplest
```{r}
# mydata <- list(n_coef = 10,
#                y = a, 
#                sigma = b)
# 
# fit <- stan(file = 'simplest.stan', data = mydata, 
#             iter = 1000, chains = 4)
```


Failed attempt to run hierarchical model
```{r}
# mydata <- list(n_coef = 10,
#                n_studies = n_studies, # number of studies
#                y = coef_array, 
#                sigma = covar_array)
# 
# # fit <- stan(file = 'nested.stan', data = mydata, 
# #             iter = 2000, chains = 4)
# fit <- stan(fit = fit, data = mydata, 
#             iter = 1800, chains = 4, 800)
```

Model estimating each interaction separately
```{r}
mydata <- list(J = n_studies,
               y = coef_vect,
               sigma = var_vect^0.5)

fit <- stan(file = 'nested_separate.stan', data = mydata,
            iter = 1000, chains = 4)
# fit <- stan(fit = fit, data = mydata,
#             iter = 2000, chains = 4)
```


```{r}
shinystan::launch_shinystan(fit)
 
```
